+++ CHANGELOG.md
@@ -1,9 +1,20 @@
 # Change Log
+### v0.3.1  
+**Fixed** 
+
+* Fixed [bug](https://github.com/shaharkazaz/vscode-git-merger/issues/10)  
+
+**Updated** 
+
+* Changed "GitMerger: Drop stash" into: "GitMerger: Delete stash"
+* Removed "GitMerger: Pop"
+
 ### v0.3.0
 **Added**  
 
 * "GitMerger: Stash", "GitMerger: Unstash", "GitMerger: Drop stash", "GitMerger: Pop"  & "GitMerger: Clear stash" commands.
-* Information Log 'Git merger info'(sperated for the error log)
+* Information Log 'Git merger info'(sperated for the error log)  
+
 ### v0.2.5  
 **Fixed**